Added ForgetPassword feature with BCryptPasswordEncoder

+++ src/main/java/com/xe/repo/ExchangeRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ExchangeRepository extends JpaRepository<Exchange,Integer> {
+public interface ExchangeRepository extends JpaRepository<Exchange,Long> {
 }
